refactor(cli/migrate): rename file to match codebase conventions
