+++ packages/remix-dev/cli/migrate/index.ts
@@ -1,2 +1,2 @@
-export * from "./input-resolve";
+export * from "./resolveInput";
 export * from "./run";
